regulators: Retry failed sensor monitoring
If a failure occurs while trying to read voltage regulator sensors,
retry the operation 5 times before logging an error.

This provides "de-glitching" to ignore transient hardware problems.

+++ phosphor-regulators/src/sensor_monitoring.hpp
@@ -117,9 +117,9 @@ class SensorMonitoring
     ErrorHistory errorHistory{};
 
     /**
-     * Number of errors that have occurred.
+     * Number of consecutive errors that have occurred.
      */
-    unsigned int errorCount{0};
+    unsigned short errorCount{0};
 };
 
 } // namespace phosphor::power::regulators
